Fix build error for editing "docId" in "docusaurus.config.js"
Need to make sure that the "docId" value matches an existing Markdown file
in the docs directory. If the docId value changes to a non-existent
file, Docusaurus will not be able to find it during the build process
and will generate an error. To fix this issue, creating a new
Markdown file in the docs directory with the same name as the docId
value wanted to be used is needed.